Rename to dsr_policy.py
