+++ scripts/setup/dsr_policy.py
